# 187 - Optimize pytorch weight loading on startup + dependencies update