# Description

Fixes PyTorch auto-downloading models on node-start even if its not used (causes significant delay until ready to process).
Dependencies updated to fix building issues (b5 update) and have newest versions of PyTorch/Ultralytics and other libraries.

Fixes #187 

## Type of change

- Bug fix (non-breaking change which fixes an issue)

# Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] My changes generate no new warnings