@MaxJa4 2024-02-03
Delay reduced to ~0.5s on my machine. Won't be at university until Thursday due to work - would appreciate a quick test of this fix on the lab PCs to see if it works :) Thanks