@samuelkuehnel 2024-02-05
Here is a short video from the delay, it is about 1-2 seconds when looking at the ros time.
[delay.webm](https://github.com/una-auxme/paf23/assets/51356601/7f727aaa-cc45-435d-bcee-c37ab4a3c0ce)