@MaxJa4 2024-02-05
@samuelkuehnel Thanks for testing. I saw that with the recent additions in the VisionNode by the LIDAR/distance features by Leon, the overall calculation time increased. It's ~50ms (20 FPS) on my PC, looks like it's more like 200ms on the lab PCs. I can try to check if I can do anything to fix this in the VisionNode without breaking the distance/LIDAR stuff, but I doubt it.