@MaxJa4 2024-02-12
Added further measures to ensure up-to-date classification data:
- Traffic light detection is handled asynchronously now (-> independent of distance calculation)
- Switch to second best classification model (YOLO RTDETR-L instead of RTDETR-X) which ~54% faster with only slightly worse accuracy according to [Ultralytics](https://docs.ultralytics.com/models/rtdetr/#pre-trained-models)

Overall, the timings of the VisionNode changed like this (tested on a 4070 Ti):

Before:
- Classification: ~50ms
- Traffic light detection & distance calc: ~22ms
- Total: ~72ms -> 13.9 FPS

Now:
- Classification: ~40ms
- Traffic light detection: asynchronously now, doesn't wait for distance calc
- Distance calc: ~15ms
- Total: ~60ms -> 16.7 FPS

Using the NAS models was not possible because of (hours of) dependency hell with Ultralytics + Super-Gradients... also converting the RT-DETR model to a way faster format like TensorRT is not loadable with Ultralytics.

**Dependencies changed, please run `b5 update`! :)**