@MaxJa4 2024-02-28
5 days ago, a new/improved version of the standard YOLO model has been released: https://github.com/WongKinYiu/yolov9
This may be a candidate to try if the above changes are not sufficient performance wise.